Revert "Add monthly menu screen"
