+++ app/src/main/java/com/example/basic/FoodMenuScreen.kt
@@ -41,6 +41,15 @@ import java.util.*
  * Simplified food menu screen inspired by the vit-student-app implementation.
  */
 
+data class Meal(
+    val name: String,
+    val startHour: Int,
+    val startMinute: Int,
+    val endHour: Int,
+    val endMinute: Int,
+    val items: List<String>
+)
+
 private val sampleMenu = listOf(
     Meal(
         "Breakfast",
