+++ app/src/main/java/com/example/basic/MainActivity.kt
@@ -64,7 +64,7 @@ fun BasicApp() {
         NavItem(R.string.more, Icons.Filled.MoreHoriz, Icons.Outlined.MoreHoriz)
     )
     var selectedIndex by remember { mutableIntStateOf(0) }
-    var foodScreen by remember { mutableStateOf(0) } // 0 - menu, 1 - summary, 2 - monthly
+    var foodScreen by remember { mutableStateOf(0) } // 0 - menu, 1 - summary
 
     MaterialTheme {
         Scaffold(
@@ -96,13 +96,10 @@ fun BasicApp() {
                     1 -> PlannerScreen()
                     2 -> AttendanceScreen()
                     3 -> {
-                        when (foodScreen) {
-                            0 -> FoodMenuScreen(
-                                onShowSummary = { foodScreen = 1 },
-                                onViewMonth = { foodScreen = 2 }
-                            )
-                            1 -> FoodSummaryScreen(onBack = { foodScreen = 0 })
-                            else -> MonthlyMenuScreen(onBack = { foodScreen = 0 })
+                        if (foodScreen == 0) {
+                            FoodMenuScreen(onShowSummary = { foodScreen = 1 })
+                        } else {
+                            FoodSummaryScreen(onBack = { foodScreen = 0 })
                         }
                     }
                     else -> Text(text = stringResource(id = items[selectedIndex].labelRes))
