--- app/src/main/java/com/example/basic/Meal.kt
@@ -1,10 +1,0 @@
-package com.example.basic
-
-data class Meal(
-    val name: String,
-    val startHour: Int,
-    val startMinute: Int,
-    val endHour: Int,
-    val endMinute: Int,
-    val items: List<String>
-)
